[modify] remove unused files, refactoring
