+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".src.main.GlobalApplication"
+        android:name=".utils.ApplicationClass"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -48,9 +48,6 @@
         <activity android:name=".src.groupMoim.MoimPersonalActivity" />
         <!--makeMoim-->
         <activity android:name=".src.makeMoim.MakeMoimActivity" />
-        <activity android:name=".src.makeMoim.SelectDateActivity" />
-        <activity android:name=".src.makeMoim.SelectTimeActivity"/>
-        <activity android:name=".src.makeMoim.SettingActivity"/>
         <!--user-->
         <activity android:name=".src.user.UserActivity"/>
         <!--viewMoim-->
