--- app/src/main/java/com/legends/moim/config/BaseFragment.kt
@@ -1,18 +1,0 @@
-package com.legends.moim.config
-
-import android.view.View
-import androidx.fragment.app.Fragment
-
-open class BaseFragment: Fragment(), View.OnClickListener, BaseDialog.BaseDialogClickListener {
-
-    override fun onClick(v: View?) {
-
-    }
-
-    override fun onOKClicked() {
-
-    }
-
-//    override fun onCancelClicked() {
-//    }
-}
