--- app/src/main/java/com/legends/moim/src/makeMoim/SelectTimeActivity.kt
@@ -1,6 +1,0 @@
-package com.legends.moim.src.makeMoim
-
-import com.legends.moim.config.BaseActivity
-
-class SelectTimeActivity: BaseActivity() {
-}
