--- app/src/main/java/com/legends/moim/src/makeMoim/SettingActivity.kt
@@ -1,6 +1,0 @@
-package com.legends.moim.src.makeMoim
-
-import com.legends.moim.config.BaseActivity
-
-class SettingActivity: BaseActivity() {
-}
