+++ app/src/main/java/com/legends/moim/utils/GlobalConstVal.kt
@@ -1,6 +1,5 @@
 package com.legends.moim.utils
 
-
 const val CHOICE_LIKE = 1       //선호
 const val CHOICE_POSSIBLE = 2   //가능(Default)
 const val CHOICE_DISLIKE = 3    //비선호
